chore: refactor functions - move to separate files
